support `@pal/system-info` module
