+++ cocos/asset/assets/texture-cube.ts
@@ -27,12 +27,12 @@ import { ccclass, serializable } from 'cc.decorator';
 import { ccwindow, cclegacy } from '@base/global';
 import { error } from '@base/debug';
 import { js } from '@base/utils';
+import { OS } from '@pal/system-info';
 import { TextureType, TextureInfo, TextureViewInfo, BufferTextureCopy } from '../../gfx';
 import { ImageAsset } from './image-asset';
 import { PresumedGFXTextureInfo, PresumedGFXTextureViewInfo, SimpleTexture } from './simple-texture';
 import { ITexture2DCreateInfo, Texture2D } from './texture-2d';
 import { sys } from '../../core';
-import { OS } from '../../../pal/system-info/enum-type';
 
 export type ITextureCubeCreateInfo = ITexture2DCreateInfo;
 /**
