+++ cocos/gfx/webgl/webgl-swapchain.ts
@@ -23,7 +23,7 @@
 */
 
 import { ALIPAY, RUNTIME_BASED, BYTEDANCE, WECHAT, LINKSURE, QTT, COCOSPLAY, HUAWEI, EDITOR, VIVO, TAOBAO, TAOBAO_MINIGAME, WECHAT_MINI_PROGRAM } from 'internal:constants';
-import { systemInfo } from 'pal/system-info';
+import { systemInfo, BrowserType, OS } from '@pal/system-info';
 import { warnID, warn, debug } from '@base/debug';
 import { WebGLCommandAllocator } from './webgl-command-allocator';
 import { WebGLStateCache } from './webgl-state-cache';
@@ -32,7 +32,6 @@ import { Format, TextureInfo, TextureFlagBit, TextureType, TextureUsageBit, Buff
 import { Swapchain } from '../base/swapchain';
 import { IWebGLExtensions, WebGLDeviceManager } from './webgl-define';
 import { macro } from '../../core';
-import { BrowserType, OS } from '../../../pal/system-info/enum-type';
 import { IWebGLBlitManager } from './webgl-gpu-objects';
 
 const eventWebGLContextLost = 'webglcontextlost';
@@ -311,7 +310,9 @@ export class WebGLSwapchain extends Swapchain {
         nullTexRegion.texSubres.layerCount = 6;
         WebGLDeviceManager.instance.copyBuffersToTexture(
             [nullTexBuff, nullTexBuff, nullTexBuff, nullTexBuff, nullTexBuff, nullTexBuff],
-            this.nullTexCube, [nullTexRegion],
+            this.nullTexCube,
+
+            [nullTexRegion],
         );
         this._blitManager = new IWebGLBlitManager();
     }
