+++ cocos/serialization/deserialize-dynamic.ts
@@ -30,10 +30,10 @@ import { cclegacy } from '@base/global';
 import { js } from '@base/utils';
 import { CCClass } from '@base/object';
 import { ENUM_TAG, BITMASK_TAG } from '@base/object/internal';
+import { Platform } from '@pal/system-info';
 import { misc, sys, CustomSerializable, DeserializationContext, deserializeTag, SerializationInput } from '../core';
 import { MissingScript } from '../misc/missing-script';
 import { Details } from './deserialize';
-import { Platform } from '../../pal/system-info/enum-type';
 import type { deserialize, CCClassConstructor } from './deserialize';
 import { CCON } from './ccon';
 import { Asset } from '../asset/assets';
