+++ cocos/ui/editbox/edit-box-impl.ts
@@ -28,6 +28,7 @@
 import { screenAdapter } from 'pal/screen-adapter';
 import { ccwindow } from '@base/global';
 import { isDescendantElementOf } from '@pal/utils';
+import { BrowserType, OS } from '@pal/system-info';
 import { BitmapFont } from '../../2d/assets';
 import { director } from '../../game/director';
 import { game } from '../../game';
@@ -39,7 +40,6 @@ import { EditBox } from './edit-box';
 import { tabIndexUtil } from './tabIndexUtil';
 import { InputFlag, InputMode, KeyboardReturnType } from './types';
 import { EditBoxImplBase } from './edit-box-impl-base';
-import { BrowserType, OS } from '../../../pal/system-info/enum-type';
 
 const ccdocument = ccwindow.document;
