+++ packages/cc-ambient-types/src/index.d.ts
@@ -15,7 +15,6 @@
 /// <reference path="./pal/input.d.ts"/>
 /// <reference path="./pal/minigame.d.ts"/>
 /// <reference path="./pal/screen-adapter.d.ts"/>
-/// <reference path="./pal/system-info.d.ts"/>
 /// <reference path="./pal/wasm.d.ts"/>
 /// <reference path="./pal/image.d.ts"/>
