+++ pal/audio/minigame/player-minigame.ts
@@ -23,13 +23,12 @@
 */
 
 import { minigame } from 'pal/minigame';
-import { systemInfo } from 'pal/system-info';
+import { systemInfo } from '@pal/system-info';
 import { TAOBAO, TAOBAO_MINIGAME, HUAWEI, VIVO, OPPO } from 'internal:constants';
 import { EventTarget } from '@base/event';
 import { AudioEvent, AudioPCMDataView, AudioState, AudioType } from '../type';
 import { clamp, clamp01 } from '../../../cocos/core';
 import { enqueueOperation, OperationInfo, OperationQueueable } from '../operation-queue';
-import { OS } from '../../system-info/enum-type';
 
 export class OneShotAudioMinigame {
     private _innerAudioContext: InnerAudioContext;
