+++ pal/audio/minigame/player-web.ts
@@ -23,10 +23,10 @@
 */
 
 import { minigame } from 'pal/minigame';
-import { systemInfo } from 'pal/system-info';
+import { systemInfo } from '@pal/system-info';
 import * as debug from '@base/debug';
-import { clamp01 } from '../../../cocos/core';
 import { EventTarget } from '@base/event';
+import { clamp01 } from '../../../cocos/core';
 import { audioBufferManager } from '../audio-buffer-manager';
 import AudioTimer from '../audio-timer';
 import { enqueueOperation, OperationInfo, OperationQueueable } from '../operation-queue';
