+++ pal/audio/native/player.ts
@@ -22,13 +22,12 @@
  THE SOFTWARE.
 */
 
-import { systemInfo } from 'pal/system-info';
+import { systemInfo, Platform } from '@pal/system-info';
 import { cclegacy } from '@base/global';
 import { EventTarget } from '@base/event';
 import { AudioType, AudioState, AudioEvent, AudioPCMDataView, AudioBufferView, AudioLoadOptions } from '../type';
 import { clamp01 } from '../../../cocos/core';
 import { enqueueOperation, OperationInfo, OperationQueueable } from '../operation-queue';
-import { Platform } from '../../system-info/enum-type';
 import { Game, game } from '../../../cocos/game';
 
 const urlCount: Record<string, number> = {};
