+++ pal/input/minigame/touch-input.ts
@@ -25,15 +25,13 @@
 import { TouchCallback } from 'pal/input';
 import { minigame } from 'pal/minigame';
 import { screenAdapter } from 'pal/screen-adapter';
-import { systemInfo } from 'pal/system-info';
-import { ALIPAY, VIVO } from 'internal:constants';
-import { Size, Vec2 } from '../../../cocos/core/math';
+import { systemInfo, Feature } from '@pal/system-info';
 import { EventTarget } from '@base/event';
+import { Size, Vec2 } from '../../../cocos/core/math';
 import { EventTouch, Touch } from '../../../cocos/input/types';
 import { touchManager } from '../touch-manager';
 import { macro } from '../../../cocos/core/platform/macro';
 import { InputEventType } from '../../../cocos/input/types/event-enum';
-import { Feature } from '../../system-info/enum-type';
 
 export class TouchInputSource {
     private _eventTarget: EventTarget = new EventTarget();
