+++ pal/input/native/gamepad-input.ts
@@ -23,10 +23,9 @@
 */
 
 import { js } from '@base/utils';
-import { systemInfo } from 'pal/system-info';
+import { systemInfo, Feature } from '@pal/system-info';
 import { EventTarget } from '@base/event';
 import { InputEventType } from '../../../cocos/input/types/event-enum';
-import { Feature } from '../../system-info/enum-type';
 import { EventGamepad } from '../../../cocos/input/types';
 import { InputSourceButton, InputSourceDpad, InputSourceOrientation, InputSourcePosition, InputSourceStick } from '../input-source';
 import { Quat, Vec3 } from '../../../cocos/core';
@@ -322,8 +321,7 @@ export class GamepadInputDevice {
                 axisValue = this._axisToButtons(value);
                 break;
             default:
-                if (code === 7) { this._nativeButtonState[Button.BUTTON_L2] = value; }
-                else if (code === 8) { this._nativeButtonState[Button.BUTTON_R2] = value; }
+                if (code === 7) { this._nativeButtonState[Button.BUTTON_L2] = value; } else if (code === 8) { this._nativeButtonState[Button.BUTTON_R2] = value; }
                 break;
             }
             if (negativeButton && positiveButton && axisValue) {
