+++ pal/input/web/accelerometer-input.ts
@@ -23,11 +23,10 @@
 */
 
 import { AccelerometerCallback } from 'pal/input';
-import { systemInfo } from 'pal/system-info';
+import { systemInfo, BrowserType, OS } from '@pal/system-info';
 import { screenAdapter } from 'pal/screen-adapter';
 import { warn } from '@base/debug';
 import { EventTarget } from '@base/event';
-import { BrowserType, OS } from '../../system-info/enum-type';
 import { EventAcceleration, Acceleration } from '../../../cocos/input/types';
 import { InputEventType } from '../../../cocos/input/types/event-enum';
