+++ pal/input/web/mouse-input.ts
@@ -24,13 +24,12 @@
 
 import { EDITOR, TEST } from 'internal:constants';
 import { MouseCallback } from 'pal/input';
-import { systemInfo } from 'pal/system-info';
+import { systemInfo, Feature } from '@pal/system-info';
 import { screenAdapter } from 'pal/screen-adapter';
-import { EventMouse } from '../../../cocos/input/types';
 import { EventTarget } from '@base/event';
+import { EventMouse } from '../../../cocos/input/types';
 import { Rect, Vec2 } from '../../../cocos/core/math';
 import { InputEventType } from '../../../cocos/input/types/event-enum';
-import { Feature } from '../../system-info/enum-type';
 
 export class MouseInputSource {
     private _canvas?: HTMLCanvasElement;
