+++ pal/input/web/touch-input.ts
@@ -24,16 +24,15 @@
 
 import { TouchCallback } from 'pal/input';
 import { EDITOR, TEST } from 'internal:constants';
-import { systemInfo } from 'pal/system-info';
+import { systemInfo, Feature } from '@pal/system-info';
 import { screenAdapter } from 'pal/screen-adapter';
 import { warn } from '@base/debug';
-import { Rect, Vec2 } from '../../../cocos/core/math';
 import { EventTarget } from '@base/event';
+import { Rect, Vec2 } from '../../../cocos/core/math';
 import { Touch, EventTouch } from '../../../cocos/input/types';
 import { touchManager } from '../touch-manager';
 import { macro } from '../../../cocos/core/platform/macro';
 import { InputEventType } from '../../../cocos/input/types/event-enum';
-import { Feature } from '../../system-info/enum-type';
 
 export class TouchInputSource {
     private _canvas?: HTMLCanvasElement;
