+++ pal/minigame/taobao_minigame.ts
@@ -25,8 +25,8 @@
 import { IMiniGame, SystemInfo } from 'pal/minigame';
 import { cloneObject, createInnerAudioContextPolyfill, versionCompare, checkPalIntegrity, withImpl } from '@pal/utils';
 import { error, warn } from '@base/debug';
+import { Language } from '@pal/system-info';
 import { Orientation } from '../screen-adapter/enum-type';
-import { Language } from '../system-info/enum-type';
 
 //taobao IDE language   ("Chinese")
 //taobao phone language (Andrond: "cn", iPad: 'zh_CN')
