+++ pal/screen-adapter/minigame/screen-adapter.ts
@@ -25,12 +25,11 @@
 import { ALIPAY, BYTEDANCE, COCOSPLAY, VIVO } from 'internal:constants';
 import { minigame } from 'pal/minigame';
 import { ConfigOrientation, IScreenOptions, SafeAreaEdge } from 'pal/screen-adapter';
-import { systemInfo } from 'pal/system-info';
+import { systemInfo, OS } from '@pal/system-info';
 import { checkPalIntegrity, withImpl } from '@pal/utils';
 import { error, warn, warnID } from '@base/debug';
 import { EventTarget } from '@base/event';
 import { Size } from '../../../cocos/core/math';
-import { OS } from '../../system-info/enum-type';
 import { Orientation } from '../enum-type';
 
 declare const my: any;
