+++ pal/screen-adapter/web/screen-adapter.ts
@@ -24,7 +24,7 @@
 
 import { EDITOR, TEST } from 'internal:constants';
 import { ConfigOrientation, IScreenOptions, SafeAreaEdge } from 'pal/screen-adapter';
-import { systemInfo } from 'pal/system-info';
+import { systemInfo } from '@pal/system-info';
 import { checkPalIntegrity, withImpl } from '@pal/utils';
 import { cclegacy } from '@base/global';
 import { warnID } from '@base/debug';
