+++ pal/system-info/src/enum-type/browser-type.ts
