+++ pal/system-info/src/enum-type/event.ts
