+++ pal/system-info/src/enum-type/feature.ts
