+++ pal/system-info/src/enum-type/index.ts
@@ -22,6 +22,8 @@
  THE SOFTWARE.
 */
 
+import { Feature } from './feature';
+
 export * from './browser-type';
 export * from './event';
 export * from './language';
@@ -29,3 +31,7 @@ export * from './network-type';
 export * from './operating-system';
 export * from './platform';
 export * from './feature';
+
+export type IFeatureMap = {
+    [feature in Feature]: boolean;
+};
