+++ pal/system-info/src/enum-type/language.ts
