+++ pal/system-info/src/enum-type/network-type.ts
