+++ pal/system-info/src/enum-type/operating-system.ts
