+++ pal/system-info/src/enum-type/platform.ts
