+++ pal/system-info/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "outDir": "./lib/",
+    "rootDir": "./src/",
+    "baseUrl": "./"
+  },
+  "include": [
+      "src/**/*.ts"
+  ]
+}
