+++ tests/init.ts
@@ -38,7 +38,7 @@ jest.mock(
 );
 
 jest.mock(
-    'pal/system-info',
+    '@pal/system-info',
     () => jest.requireActual('../pal/system-info/web/system-info'),
     { virtual: true, },
 );
